requirements-file.txt
streamlit==1.36.0
pillow==10.2.0
pandas==2.2.0
numpy==1.26.3
